allow unions in TypeIs
Summary:
The logic for TypeIs should mirror isinstance, but there are a few differences. For example: `TypeIs[A | B]` is allowed, but `isinstance(x, A|B)` is not. Instead, `isinstance` has special handling to treat the second argument as a union when it's a tuple of types, which `TypeIs` doesn't have.

The end result is that we can just directly call intersect and subtract for TypeIs.

Reviewed By: ndmitchell

Differential Revision: D71335226

fbshipit-source-id: 11bfed0bedab13a321a1b000f1418b1cd15a5883