track when variables are unbound
Summary:
When a name is deleted with `del` or if a variable is annotated but not initialized, there's no value associated with that name anymore so we shouldn't be able to use it in an expression.

To do this, we introduce a new `FlowStyle::Unbound` that gets added when we `del` a variable. This gets cleared the next time the variable is assigned to during `bind_assign` which makes the flow style `None`.

Is tracking this via the flow style a sensible approach?

Reviewed By: ndmitchell

Differential Revision: D71330226

fbshipit-source-id: f79ff810cc4aff7fb2579848740cbcfe96e6cf50