Implement basic `__getattr__` support
Summary:
nouserattributes

This diff attempts to add `__getattr__()` support in response to [this Github issue](https://github.com/facebook/pyrefly/issues/23).

The basic behavior I'm trying to implement is this: when we attempt to perform an attribute lookup on base `b` with name `n` and the lookup failed (i.e. resolving the type of `b.n` where `n` is not explicitly defined anywhere on `b` or its superclasses), we attempt to perform a second attribute lookup on `__getattr__` -- if this result hits, then we try to resolve the type of `b.n` as if we are resolving the type of `b.__getattr__("n")`.

If we look at it, this process is very similar to how descriptor works: instead of resolving the type `b.n` directly, we "delegate" the type to another function call later. In the case of descriptor, the function call is `__get__` or `__set__`. In the case of getattr lookup, the function call is `__getattr__`.

Based on such observation, my implementation of `__getattr__` closely followed the implementation of descriptors: I added another variant to `LookupResult::Found` called `GetAttr`, preserve all info necessary to resolve the type of the `__getattr__` function call, and then resolve the function call in `resolve_get_access()`.

Attribute access is a rather low-level operation which is likely to be depended on by a large part of the type checker. Injecting a hook for __getattr__ can end up interacting with many other features. I'm trying to follow the path of least resistance here and do the simplest thing when interaction occurs. But we may need to revisit many of the decisions here (hopefully ask for a spec clarification) in the future when more feature requests come in.

Reviewed By: ndmitchell

Differential Revision: D71367405

fbshipit-source-id: 14aa412c6aeadb0e6984d786f7797cd7b7bf9cd8