Count the frequency of decorators and save to directory
Summary:
This diff counts the number of times each decorator is used / called.

This would help to decide which decorators are uninteresting. Then we can
ignore them to avoid unnecessarily increasing the trace lengths.

Saving these decorator counts to a directory would be useful in the long run
(as opposed to a one-off thing), because then we could easily check the
decorator counts from any run and evaluate which new decorators are
uninteresting.

Reviewed By: alexkassil

Differential Revision: D71403058

fbshipit-source-id: 3581946b61a144b56cc91affb2c79a158a368f08