Add a test for building higher order call graphs for decorated targets
Summary:
A later diff would fix this test. The problem is that the dependency graph does
not include the decorated targets, because it is built from the original call
graphs (which do not include the decorated targets).

As a result, when analyzing
```
def foo@decorated():
  return decorator(foo)
```
the global fixpoint would not reanalyze `foo@decorated` when the model of
`decorator` changes. This is clearly wrong.

Reviewed By: arthaud

Differential Revision: D71415598

fbshipit-source-id: 896bcb280afa272cd7533e6405bc90c4676676f3