Fix missing filename in models
Summary:
# Problem
In D71134905, we removed the module (which is used to find the filename) from the `CallInfo` part of the taint representation, assuming the module/filename was already specified at the root of the model.
Unfortunately, it looks like this is not always the case. SAPP started silently erroring when parsing our taint output, for instance:

```
2025-03-18 08:07:24,265 [ERROR] Unable to parse trace for the following:
{'callable': 'Obj{xxx[redacted]}',
 'model_generators': ['[redacted]'],
 'sources': [{'port': 'result',
              'taint': [{'declaration': None,
                         'kinds': [{'kind': 'UserControlled'}]}]}]}
```

There are two problems:
* (1) Models for attributes and global variables don't have a filename. This is expected, since those models can't have sources or sinks with a `CallInfo.Origin` or `CallInfo.Declaration` (only `CallInfo.Declaration`). The problem is that the parser doesn't know that.
* (2) Models for functions or methods sometimes don't have a filename.

# This diff
This diff solves problem (2), by forcing all models for functions and methods to have a filename, even if unknown (i.e `*`).

There seems to be another problem, way too many functions get assigned the unknown filename. This will be fixed in a follow-up.

Reviewed By: tianhan0

Differential Revision: D71466272

fbshipit-source-id: b354ec77d153504e95a16cf6f9c6016120cfb591