Create FilteredGlobs FileList type
Summary:
In upcoming diffs, I'm going to add `project_excludes`, which is also a `Globs`. We want to use this to adjust how we're enumerating files to type check, so we can filter out files that should be excluded. We'll do this by:
- exposing `Globs::matches()`
- Creating a new `PatternGlobs`, which consists of include and exclude globs (this diff)
- taking in new flags and config items for the `project_excludes`
- Passing the `PatternGlobs` as a `FileList` to places where we need a file list

This diff creates a new `FilteredGlobs` struct, which we can pass around as a `FileList` and can filter paths returned by `includes` by matching on `excludes`.

Reviewed By: grievejia

Differential Revision: D71511508

fbshipit-source-id: 5de0c81beb8a1113649c96be01fc6107b1caaa2b