Make glob package accessible
Summary:
In upcoming diffs, I'm going to add `project_excludes`, which is also a `Globs`. We want to use this to adjust how we're enumerating files to type check, so we can filter out files that should be excluded. We'll do this by:
- exposing `Globs::matches()`
- Creating a new `PatternGlobs`, which consists of include and exclude globs
- taking in new flags and config items for the `project_excludes`
- Passing the `PatternGlobs` as a `FileList` to places where we need a file list

This diff exposes the `globs` module, instead of only exporting `globs::Globs`

Reviewed By: grievejia

Differential Revision: D71511509

fbshipit-source-id: b89f6b63b5c4af3de9ff5d6544c867345ead6cda