Add a test case to `callGraphFixpointTest.ml`
Summary:
We currently do not handle the pattern of descriptor protocol (that are used as
decorators), because this requires handling the propagation of functions into
and from object attributes. As a result, the returned callables are empty for
any decorator that uses the descriptor protocol, or more generally, any
decorator of a class type that stores the decorated functions into object
attributes.

This diff adds a test case that catches this pattern.

Reviewed By: arthaud

Differential Revision: D71487956

fbshipit-source-id: 279eda4eabe79ac33b4b5c5cdf78a689c0d82596