handle nonlocal
Summary:
Changing the scope/lookup behavior has a lot of interactions with other logic so I ended up doing it in one go to avoid breaking any tests. Looking for comments on the overall approach + LMK if you want me to try and split it.

There are 3 main changes:
- LookupKind::Nonlocal figures out what the variable in a nonlocal declaration refers to
- other lookup types will check enclosing scopes if there's a nonlocal declaration in the current scope
- when searching for annotations for an assignment, check enclosing scopes if there's a nonlocal declaration in the current scope (previously, it only checked the current static scope)

The implementation of the third bit is a little half-baked but it does the job for now. I don't know how it behaves outside of the happy path, so it probably needs revisiting later.

Reviewed By: ndmitchell

Differential Revision: D71438790

fbshipit-source-id: 40d7162e9643ea31bf30b52578c24986e3af2baf