Turn `get_handles()` into a struct
Summary:
aliens_structs

Previously, we rely on the `get_handles()` function to "transform" a list of file paths into a list of handles.

A function works well for one-shot full checks. But in watch mode, the transformation is unfortunately unstable across incremental updates. We need the transformation itself to keep some internal states that can update accordingly when the file watcher notifies it of certain filesystem events. For that purpose, we want to turn `get_handles()` from a function, which cannot store internal states, into a struct which can do that.

I call this struct `Handles` (open to better naming) and for now we'll make things simple by only exposing 2 public APIs: one is to construct a `Handle` from a list of files + a config finder, and one is to grab all handles that needs to be checked at the moment. We'll extend the struct with a 3rd API which support incremental update in D71522935.

Reviewed By: stroxler

Differential Revision: D71522936

fbshipit-source-id: 86c01a0612306e36bd4c4b7eee676c50516ff2c5