Track which path maps to which loader during handle creation
Summary:
refactor

This diff attempts to rewrite the `Handles` class, so we could keep track of how many loaders we have created and which files are using which loaders. No functional changes are intended -- the original semantics of `Handles::new()` and `Handles::all()` should be preserved. I'm just trying to use better data structures to store the internal states so incrementally handling file creation and deletion events can be coded more easily(see D71522935).

In a world where files to check remain stable and all files to check share exactly the same `Loader`, it's very easy to handle those events: simply call `State::invalidate_find()` which takes a reference to that one single shared loader, and we are done. But two features were added after D69241375 which makes invalidation on file creation/deletion trickier:
- First, we now support file globbing. This means that file creation/removal may now affect what handles we want to send to `State.run()`. For example, if the user says `pyrefly check foo/**/*.py`, and a new file `foo/bar.py` is added, then we can't pass the same set of `handles` we've calculated before to `State` -- we've got to update the handle list and include additional corresponding entry on `foo/bar.py`. Same for removal.
- Second, we now support per-file config finding. This means that different files may correspond to different `Loader` when being converted into handles. Say we have two files `P` and `Q` both have their own search paths and therefore they use distinct loaders -- when we receive a file removal event on `P`, then ideally we only want to call invalidate find on the loader of `P`, not that of `Q`.

The first item requires that our `Handles` struct need to somehow keep track of the entire set of handles are currently being checked in its internal state, so we would have an easier time updating them when files get added/removed.

The second item requires that our `Handles` struct need to somehow keep track of which file is using which loader, so we would have an easier time knowing which loaders to invalidate when files get added/removed.

... And that's exactly what this diff does: it changes the internal state of `Handles` from a `Vec<(PathBuf, ConfigFile)>` field (which is not easy to update incrementally) to a `Map<PathBuf, (Loader, /* some extra data */)>`  field. The map is a better structure for incremental update because (1) it's easy to turn it into a list of handles: just iterate over the entire map and use the keys & values to construct handles, (2) it's easy to maintain on file addition/removal since we could just turn those events straightforwardly into map insertion/removal, (3) it's easy to figure out which loaders get affected from file addition/removal events: just perform a map lookup on those files and collect the results. We also need a `Map<LoaderInput, Loader>` on the side so we could deduplicate the `Loader` according to the search paths. (Note that the structure also provides a natural place to deduplicate `ConfigFile` if we want to pursue that direction -- cc. connernilsen).

Reviewed By: connernilsen

Differential Revision: D71522938

fbshipit-source-id: 16ad55b9d5c27a4333c2708bee71aa2480a00482