Wrap CLI config override logic into another `config_reader`
Summary:
justmovingthingsaround

In `CheckArgs::get_handles()`, we find the config of a given file `F` by first invoke `config_finder(F)` and then override the found config with the CLI args specified in `CheckArgs`. This means that the `get_handles()` function must take a `CheckArgs` argument (in the current implementation, that's the `self` argument).

I'm rewriting the code so the `get_handles()` function no longer need to be aware of `CheckArgs` -- this is achieved by hiding the CLI config override logic behind the `config_finder` closure itself. This enables further refactoring in D71522937.. No functional changes is intended here.

Reviewed By: stroxler

Differential Revision: D71522939

fbshipit-source-id: 96b4544bba0d7567bb8bd85e0d0a92aa2fff10f1