Get windows paths working again for configs
Summary:
X-link: https://github.com/facebook/pyrefly/pull/30

We do not handle paths well on Windows, especially if a user passes / in their paths through CLI or config (this is due to how path libraries are implemented in rust). To fix this, we need to sanitize all the paths we get from CLI or configs, either by wrapping everything in a `Platform`-prefixed newtype that we can use as a sanity check (and in construction, fix up the newtype ourselves), or by making a new trait that can do the conversion for us, and calling it everywhere we get user data.

For now, to get things working again, let's just sanitize that path when we're working in the config, and assume our users will be nice to us by keeping their configs' paths platform-specific.

Reviewed By: ndmitchell

Differential Revision: D71589137

fbshipit-source-id: 5783481893c417b84717ef151257905e2643898c