force variable in tuple subtyping
Summary:
The type argument for the class representation of tuple is sometimes a variable.

We need to have special behavior for when the type argument is `Any`, since it would be equivalent `tuple[Any, ...]`.

However, we can't use is_subset/is_equal to compare the type argument with `Any`, since it will always be true regardless of what the type argument is due to Any's special semantics.

To get around that, we force the variable first before we compare with `Any`.

Reviewed By: ndmitchell

Differential Revision: D71646611

fbshipit-source-id: 744525d455e237bed41a88195929129f47cf548e