fix tuple ClassType display
Summary:
Tuples have a Type::ClassType representation as well as their usual Type::Tuple. The former is used when we construct a tuple from an iterable using the constructor, and when we call methods on the latter.

Previously, ClassType(tuple, [int]) prints as tuple[int] when it's actually equivalent to tuple[int, ...], leading to confusion

Reviewed By: ndmitchell

Differential Revision: D71737616

fbshipit-source-id: 44123937876994e2d20e7f0ebc6391127a265037