Invalidate all loaders on any file addition/removal
Summary:
cat_shame

D71522935 made it such that when a file-to-check `F` is added/removed during incremental check, we invalidate the loaders used to load `F`. If `F` is not checked, then we don't invalidate any loader.

But that's actually incorrect! Even if `F` is not type checked, it could still serve as a dependency to one or more files-to-check, whose loader then in turn needs to be invalidated.

Figuring out which loader is affected when `F` serves as a dependency is actually tricky to do outside of the State code. Therefore, I believe the right thing to do currently is to `invalidate_find()` on all loaders on any file add/remove. It may invalidate more than we need, but it's the safe thing to do for now.

Reviewed By: stroxler

Differential Revision: D71754304

fbshipit-source-id: dbc5944ca52f612c782816929fa306abba5f7916