Introduce `FileList::covers()`
Summary:
For incremental checks, the set of files on which we need to report type errors may change as the filesystem gets updated.

what files the file watcher is watching for changes may not always match what files need to be type checked. For example, if the user does
```
pyrefly check 'foo/*.py' --watch
```
We collect what files to check when the type checker is started -- say we find two files `/test/foo/bar.py` and `/test/foo/baz.py`. Later on a user may choose to either throw new py files into `/test/foo/` or remove pre-existing py files from `/test/foo/`. In those cases, we'd want to adjust the list of files to check accordingly so we know on which files we should query for type errors in the next incremental check.

In practice, the file watcher may choose to watch more files than what's listed as files to check. For example, we may choose to watch the entire directory `/test/` while only checking for files under `/test/foo/`. In that case, what's also necessary is to make sure that whenever a file watcher hands us a path `P`, we know whether this `P` is "covered" by the files-to-check list. Again in our running example, if `P` is `/test/foo/qux.py`, then it should be "covered" and we should have `Handles` act on it. But if `P` is `/test/derp/qux.py`, then we should ignore it when updating the files-to-check list. Note that checking if a file is "covered" or not does not depend on the actual existence of `P` -- `P` may actually be non-existent (e.g. if it comes from a file removal event) but we still need an answer from the "is covered" check.

This diff adds a new interface `FileList::covers()` which perform the aforementioned check. For OSS `FileList`, the `covers()` operation is pretty much just the `matches()` operation except that I want a `bool` instead of a `Result<bool>`. I'll fill in the implementation on the IG `FileList` in D71590616.

Reviewed By: connernilsen

Differential Revision: D71590618

fbshipit-source-id: 83985fa90c3af8d4b54144a99b0198fc6f6059ff