Account for suppressed errors when exiting
Summary:
If there are any errors and they're all suppressed, we report that the check succeeded, then the info line says there were X errors, and then pyrefly exits with a non-zero code. This is confusing, particularly in CI.

Instead, show the suppressed count (if there are any) and factor that in when determining the return code.

Reviewed By: ndmitchell, grievejia

Differential Revision: D71777298

fbshipit-source-id: 36e81fd54935acbe03ef8f201cf9d43f5cf19cc4