basic go-to definition /. hover on modules
Summary:
Use imports from ast::imports to allow hover / go-to definition on them. Since they are not visited in ruff's ast visit, we must find them manually (this diff uses `Ast::imports`). An alternate approach would be finding them on the same visit as identifiers (D71527463) but I think this is cleaner.

note: this does not support go-to definition parts of the import. for example, clicking `foo` in `import foo.bar.baz` taking you to `foo/__init__.py`.

Reviewed By: ndmitchell

Differential Revision: D71776221

fbshipit-source-id: 43ab12116149669f193a1685dfdcc3ec64372b70