Delete AnnotationTarget::name()
Summary: Having a "name" method on AnnotationTarget is a little misleading because the name doesn't mean the same thing in every variant (e.g., AnnotationTarget::Return stores the name of the function). The method is used only in one place, so we can easily inline it. We also know that a Binding::IterableValue can't appear in a function signature.

Reviewed By: Solumin

Differential Revision: D71845946

fbshipit-source-id: 0ab20c3befa8c8ea24453ad17aad0d3bc8206392