Remove `State::print_errors()`
Summary:
We have `State.collect_errors()` and then `State.print_errors()` duplicating the logic of grabbing the errors from the state. This makes it a bit harder for me to change the relevant logic upstream.

This diff proposes that we implement `State.print_errors()` on top of what's collected in `State.collect_errors()`, so they are guaranteed to share the same error collecting logic. Once we start doing that, it becomes clear that the `print_errors()` functionality does not really depend on anything within `State`. So we could simply lift it up to elsewhere as a free function.

Reviewed By: Solumin

Differential Revision: D71840826

fbshipit-source-id: d279b6b83dbc7c566da9e7bcd6c0babea3ecfaa8