Remove `State::print_error_counts()`
Summary:
We have `State.collect_errors()` and then `State.print_error_counts()` duplicating the logic of grabbing the errors from the state. This makes it a bit harder for me to change the relevant logic upstream.

This diff proposes that we implement `State.print_error_counts()` on top of what's collected in `State.collect_errors()`, so they are guaranteed to share the same error collecting logic. Once we start doing that, it becomes clear that the `print_error_counts()` functionality does not really depend on anything within `State`. So we could simply lift it up to elsewhere as a free function.

Reviewed By: stroxler

Differential Revision: D71851362

fbshipit-source-id: 513524caaf00bf486178372593a4f03c8f8c43a2