Create new types for error configuration
Summary:
This diff creates the `ErrorConfig` and `ErrorConfigs` types and replaces the type used for errors in `ConfigFile`.

**Current plan**

My thoughts right now on how to do error filtering are:
1. add `ErrorConfig` and `ErrorConfigs` (this diff)
2. update the function interfaces to pass these args through or get an `ErrorConfig` from an `ErrorConfigs` and pass that through
2. add error filtering to the relevant functions in `ErrorCollector`
3. get `ErrorConfigs` when parsing `ConfigFile`s and pass those through

After talking to grievejia and ndmitchell, it seems like passing an `ErrorConfig` down when getting errors is the right way to go over giving an `ErrorCollector` an `ErrorConfig`, given that they seem to be created all over the place in an ad-hoc manner, and having the `ErrorConfig` be a part of that would be kind of tough.

**The new types**

The `ErrorConfig` type is a `HashMap<ErrorKind, bool>`, which we will be able to use to determine if a given error kind should be enabled/disabled. If the value is not present, we can assume it should be enabled. I may end up changing it to an `Arc<...>` version to cut down on extra copying. I think this early implementation of `ErrorConfig` shouldn't be too controversial.

What might be controversial though, are `ErrorConfigs`. The `ErrorConfigs` is a `HashMap<ModulePath, ErrorConfig>`. I'm using `ModulePath`, since that's the only identifying type we have access to right now in `ErrorCollector`. We need a way to identify the right `ErrorConfig` to use for the given `ErrorCollector` for the `ErrorCollector::get_errors_per_file()` and `ErrorCollector::count_error_kinds()` functions, where we're iterating over a list/iterator of `ErrorCollector`s for different modules.

Reviewed By: stroxler, ndmitchell

Differential Revision: D71823279

fbshipit-source-id: b885433e1d748d4e520fc8d2f6090e24d60f52e3