Create `ErrorConfig`s when registering file `Handle`s
Summary:
This diff handles getting an `ErrorConfig` from the parsed `ConfigFile` when processing files to type check in `Handles`, and exposes them through an `error_configs()` method.

We also have cleanup for when a file is deleted to remove the file and its error config from the `module_to_error_config` mapping.

We need to do this where we have access to the `ConfigFile`s, so right now the obvious choice is to have this functionality in `Handles`, but it still feels a little weird that it's there. If we had `config_finder` doing caching and returning an `Arc<ConfigFile>`, then it might make more sense to do this elsewhere in `run_inner()` by getting all the `ConfigFiles` again, but to make sure we're not doing a ton of extra processing, let's just leave it in `Handles` until that's ready.

**Current plan**

My thoughts right now on how to do error filtering are:
1. add `ErrorConfig` and `ErrorConfigs`
2. update the function interfaces to pass these args through or get an `ErrorConfig` from an `ErrorConfigs` and pass that through
2. add error filtering to the relevant functions in `ErrorCollector`
3. get `ErrorConfig`s when parsing `ConfigFile`s (this diff)
4. get `ErrorConfigs` from `Handles` and pass them to `run_inner()`

Reviewed By: ndmitchell

Differential Revision: D71836644

fbshipit-source-id: 01c6a4581c9a4f2d6a50a25647664ee66c20c392