Remove `State::count_errors()` and `State::count_suppressed_errors()`
Summary:
We have `State.collect_errors()` and then `State.count_errors()` (and `count_suppressed_errors()`) duplicating the logic of grabbing the errors from the state. This makes it a bit harder for me to change the relevant logic upstream.

This diff proposes that we implement `State.count_errors()` on top of what's collected in `State.collect_errors()`, so they are guaranteed to share the same error collecting logic. It is made possible with D71854831, where we could count errors outside of the state code.  Once we start doing that, it becomes clear that the `count_errors()` functionality does not really depend on anything within `State`. So we could simply inline the logic.

Reviewed By: ndmitchell

Differential Revision: D71851364

fbshipit-source-id: b61a0a7037ffd0115069485502de55e8baecef4d