Merge `ModuleErrors:iter()` and `ModuleErrors::iter_all()`
Summary: With D71854831, we now always return all errors from the error collectors. There's no longer any need for `ModuleErrors` to differentiate between iterator than include suppressed errors, and the one that does not include them. The only pre-existing caller of the non-inclusive iterator is `Display::fmt()`, but I think it's safe to change that since the display is not currently used in prod and is probably there for just debugging purposes.

Reviewed By: ndmitchell, kinto0

Differential Revision: D71937163

fbshipit-source-id: 688582d48d733bf3cf753728f6fc7c511a27e4e2