Do some clone tetris
Summary:
In order to deal with the possibility that a class field has an inferred value coming
from an implicit method-level definition that uses an invalid type var, as in
```
class C:
    def __init__[T](self, x: T): self.x = x
```
I am going to need the ability to synthesize a sanitized type where the type variables
have been eliminated.

That means I need to own `ty` in most of the body of `calculate_class_field` instead of
just holding a reference.

It turns out this has no perf implications: we already had to clone `ty` eventually in
every code path, we were just deferring it to the very end. Doing it earlier makes the
present refactor - and any future changes where we realize the type needs sanitization
rather than just checking - possible.

Reviewed By: yangdanny97

Differential Revision: D71906531

fbshipit-source-id: 6e40eed8930f3ae5ab8719bb03efd159aa6c2d4e