Add plumbing to track method where a field was inferred in bindings
Summary:
I did quite a bit of work trying to implement the check for unbound type variables
in a class field, coming from code like this that tries to initialize a field to a
value whose inferred type uses method-scoped type vars:
```
class A:
    def __init__[T](self, x: T): self.x = x
```
which is syntactically legal and works fine at runtime, but doesn't make semantic
sense in terms of static types.

My initial attempt, though, involved looking for an quantifier that isn't part of the
class type parameters or bound by a Forall in the methods.

There are two problems with this:
- First, we've made it pretty hard to find all the Foralls in a visitor,
	because instead of using Type::Forall reliably we've inlined the Forall into
	BoundMethod and Overload. My first implementation was wrong because I didn't
	realize this, and even if I get it right, it's a pretty brittle operation that
	future type system changes could easily break.
- More importantly, I realized it doesn't work anyway: that logic will incorrectly
  handle classes that only exist in a nested function and use generic function
	parameters.

For an example of the second bullet point, which is admittedly edge-case-y in
most idiomatic Python codebases,
```
def f[T](x: T) -> T:
    class A:
		    def __init__(self, x: T): self.x = x
		return A(x).x
```
is perfectly legal code as far as I know - the name `A` never leaks out of the
nested scope, and inside of that scope `T` is a concrete type.

My understanding is that some actual codebases do use the pattern of locally-scoped
classes to defer initialization costs, which means this could come up.

So, I'm now taking a second pass at this: we probably need to thread enough information
through to be able to (a) restrict the check to only method-initialized fields and
(b) explicitly look for the method-scope type vars that are known to be *bad*, instead
of trying to check all the type vars to make sure they are *good*.

This diff begins that process by threading the method name through bindings.

Reviewed By: rchen152

Differential Revision: D71906529

fbshipit-source-id: 35e89fc8981b41514e1dae7f09fda3a941364dac