Add a TODO on another fishy ReadWrite method case
Summary:
There are now two spots in `attr.rs` where we are handling methods as ReadWrite,
when they should be ReadOnly. It seems clear that some part of Pyrefly is creating
methods with the wrong visibility, and we should track it down.

Adding a TODO because it's useful to have a code pointer to the symptom, but this
isn't high-enough priority to debug right now.

Reviewed By: rchen152

Differential Revision: D72012081

fbshipit-source-id: 271bb0595ed01868fde0c4e7518bfa88aa786dd1