Add `# pyrefly-ignore`
Summary:
TLDR; our own custom error silencing comment makes it easier to onboard new projects to Pyrefly

When working on adding Pyrefly to new projects, I often have to add or remove error suppressions.

This gets difficult if we want to ensure that type ignores are not "littering" the codebase and/or MyPy is configured to ban unused ignores.

It's been argued that:

> type: ignore seems simplest. Once you have 4 different type checkers having 4 different ignore statements is going to get annoying

I strongly disagree here. While it's true it is simplest, and we should continue to support it, it makes suppressing errors in a codebase with more than one type checker more annoying.

The ideal would obviously be having one type checker per project, but while people are experimenting or in the middle of a migration, that won't always be the case. Adding a pyrefly specific comment should make it easier to convert between type checkers, both internally and externally.

As an added bonus, it would make it easier to remove unused type suppression comments with confidence, and know that we aren't removing suppressions other tools or CI jobs rely on.

Reviewed By: Solumin

Differential Revision: D72070901

fbshipit-source-id: 24c4da41b22d99c15cacd4b6aafc9bdb5430dffb