Try to fix windows path lengths for github CI
Summary:
Looks like there was a Rust update that causes paths to be too long on Windows when cloning it with Cargo. This resets the home to be `C:\cargo`, which makes it short enough to succeed again.

There's another failure happening though, but that's unrelated.

X-link: https://github.com/facebook/pyrefly/pull/57

Reviewed By: ndmitchell, kinto0

Differential Revision: D72071684

fbshipit-source-id: b0772b1f9f7f58d6ebcae42df71ae42627f8c624