Defer calculation of typed dict fields
Summary:
This diff makes the key change needed to handle cyclic typed dict definitions:
defer the calulation of their fields, so that the types don't directly try to own
the field types (which is not actually possible - the attempt to do this caused
huge slowdowns in some cases, and it also fails because at some point the recursion
has to give up).

The fields are now calculated by the `typed_dict_fields` method in `class/typed_dict.rs`
as needed.

I considered making a `Key` / `Binding` pair here, but realized it's not clear whether
there would be much savings - we have to perform substitution after the fact anyway,
which means most of the work happening on top of the cached `ClassFields`, which is
a combination of substitution and clone, is unavoidable even if we cache part of
the computation.

Without perf results that clearly indicate a problem, I don't think it's worth
the complexity of a binding.

--------

Note: there's a pretty unpleasant hack in `callable_infer_params`.

(Edit from the future: fixed in the stacked diff D72074622)

The problem is that `callable_infer_params` is a very complex method - it is over
400 lines long, and has at least 9 locals plus 8 parameter that live for the entire
function body, with around 10 more locals that live for longish blocks inside
of nested loops and matches.

All of these locals, and much of the control flow, relies heavily on never needing
to own any Types - it was designed to depend 100% on the idea that if I have
parameter types, I would never need to compute a new value in the process of analyzing
the callable parameters.

This is no longer true if typed dict fields are deferred - we will see a typed
dict in the middle of traversing parameters, and have to look up its fields which runs
a new calculation and produces an owned value.

But there's no obvious way to actually get a reference to these fields, which is what
the rest of the function depends on: the owned value lives inside of a loop, so we
cannot take a reference to it safely. I looked into trying to put the value into a
data structure that lives for the lifetime of the function, but the obvious data
structures would require a mutable reference to actually store the value, at which
point we still can't get a reference into the data structure.

After some spitballing, it seems like a stack data structure that has only a push
operation would work in principle; I think that the `append_only_vec` might implement
what we need.

But I would rather not pull in a dependency in the middle of the typed dict fields
fix, so for now I go with the simple solution of just iterating over the parameters
twice, which is both slightly less efficient and pretty confusing to read given how
huge the function is. I put a TODO in to clean it up in a stacked diff.

Reviewed By: ndmitchell

Differential Revision: D72074620

fbshipit-source-id: 0a61898efaa393dc38677bc2c21bfbfc1d94029a