Use AppendOnlyVec for heap storage
Summary:
In the previous diff, to work around the need for typed dict field
calculations to be able to produce owned values on-the-fly in the middle
of code that relies on using types by reference everywhere, I had to
implement a full extra loop over the parameters for storing typed dict fields
so that I could take references safely later.

The problem here is that any non-purpose-built data structure that lets you
push data onto the heap requires a mutable reference to do that, at which point
you cannot take references in the same block that you use the mutable API.

But in reality, it's safe in terms of actual memory usage to *append* owned
data to a vec, while taking references to only what is already there. The
`AppendOnlyVec` data structure has a restrictive API that reflects this - it allows
append on a non-mutable reference, so that we can interleave storing owned values
on the heap and taking references to them in the same block.

This allows us to iterate over the parameters just once, which is a small
efficiency win and makes the code *much* more readable, particularly given how
complex `callable_infer_params` already is.

Reviewed By: ndmitchell

Differential Revision: D72077000

fbshipit-source-id: cb1674feac21609426e76263ae7648cd059c1739