use pyrefly color palette
Summary:
This diff uses the color palette fetched from the new pyre icon, and standardizes the entire website to use the color palette. Here are the colors used:
- brown: #543D3B
- orange: #FFFFFF
- darker orange: #FF7701
- red: #FF4300

custom color:
- off-white: #efefef

Putting this up for RFC because I'm not convinced this looks better. I personally think the dark grey and yellow looks better than the brown and orange, but that is also not our color scheme.

One thing we could do is land this diff first, and then adjust the primary/secondary color after, since after this diff, we should only need to adjust the colors in one place (`custom.css`).

Reviewed By: yangdanny97

Differential Revision: D72005726

fbshipit-source-id: 32d52c9a04c6d16be94cfa94b842d5354555a77e