Rename via:lambda into via:higher-order-parameter
Summary:
The `via:lambda` feature is added on flows through our higher order parameter heuristic.
When a function is called with another function has a parameter, we assume it will be called.
For instance, `foo(bar, x)` will be treated as calling `bar(x)` implicitly. The `via:lambda` breadcrumb is added on those flows.

Since this has nothing to do with lambdas (lambda is python usually refer to a `lambda x: ..` statement), let's rename this into `via:higher-order-parameter`.

Reviewed By: tianhan0

Differential Revision: D72159957

fbshipit-source-id: 6cbd4384793fe2f7d4f48da2951d47c247fe8cec