Fix glob ignores
Summary:
I realized while testing today that the default ignores we've been using will sometimes silently ignore all files if a path is passed in with `.` as a path component in the middle of the path.

It was a sneaky problem to have, but I added tests to make sure the fix actually does what we think it should.

Interestingly, I found this while running the conformance test from `./test.py` on D72084930, and getting the error that no files were tested from the recent diff that adds validation when project excludes filters all paths.

The good news is that our conformance tests have been running, and it's just an upcoming diff that would have made them not run. Good thing our tests caught it!

Side note: removing the `.py` from the default includes pattern is okay, since we do `*.py` and `*.pyi` filtering when we call `files()`

Reviewed By: grievejia

Differential Revision: D72181187

fbshipit-source-id: c72b58d93016d58bc0a24558554065d830ff579b