Group name-oriented expressions near the top of `expr_infer_with_hint`
Summary:
This diff is just rearranging cases in the expression match.

It is debateably a small BE win - I personally think `Expr::Name` should 100% be
the first branch in our expr traversal, because these 3 lines of code - which
are where answers-side part of binding happens - explain more of Pyrefly's
architecture than any other 3 lines in our codebase.

But I'm more specifically motivated by the fact that I'm working on attribute
narrowing, and these three expressions are the first three that I'm pretty
sure will need to preserve attribute narrowing - the vast majority of expressions
perform "operations" that are type-level, and would not keep attribute narrowing.
These three do:
- `Name` and `Named` are both passing a value along as-is with no operation, which
  means they need to propagate any narrowing
- `Attribute` is indeed performing an operation (attribute lookup in Python is
  a complex operation), but this is the one operation that we actually want to
	model with attribute narrowing.

It's easier for me to hack if these cases are next to each other, and putting up
a diff now to do that will make rebase conflicts less annoying later.

Reviewed By: ndmitchell

Differential Revision: D72250256

fbshipit-source-id: 6f8c9b88432c82fe73b2c4c17d32b247bf8c63fe